+++ physics/Collider.js
@@ -2,7 +2,7 @@ import BoundingBox from "./BoundingBox.js";
 import Mesh from "../shapes/Mesh.js";
 import Vector3 from "../structs/Vector3.js";
 import MeshCuboid from "../shapes/TestShapes/MeshCuboid.js";
-import Player from "../Player.js";
+import Player from "../gameObjects/Player.js";
 
 export default class Collider {
   constructor(gameobj) {
